[π] Split constraint-set regression benchmarks into a dedicated CI shard
